move s3 object scanning & handling to own top-level module
in preparation for use by views by both main and callbacks blueprints
